re-enable Mac build (#657)
* enable mac build, enable basic OS commands as on linux

* back to intact linux stuff

* fix timing for alive status for air, ground, rs, fc and wifi card systems

* show all local IP addresses

* remove unclear define

---------

+++ app/telemetry/models/aohdsystem.cpp
@@ -458,8 +458,8 @@ void AOHDSystem::update_alive()
     // come through, not only a heartbeat
     // AIR: Quite lossy, and r.n we send about 2 to 3 telemetry packets per second
     // GROUND: OpenHD (can) send a lot of data to QOpenHD, since there are no bw constraints
-    const int tmp=m_last_message_ms;
-    if(tmp<=-1){
+    const int64_t tmp=m_last_message_ms;
+    if(tmp<=-1) {
         update_alive_status_with_hud_message(false);
         return;
     }
